add: code formatting

+++ contracts/test/MockFeeWIthdrawUniswapRouter.sol
@@ -2,14 +2,11 @@
 pragma solidity 0.8.1;
 
 contract MockFeeWithdrawUniswapRouter {
-
     function swapExactTokensForTokens(
-        uint amountIn,
-        uint amountOutMin,
+        uint256 amountIn,
+        uint256 amountOutMin,
         address[] calldata path,
         address to,
-        uint deadline
-    ) external {
-        
-    }
+        uint256 deadline
+    ) external {}
 }
