+++ contracts/test/MockUniswapV2Router02.sol
@@ -5,12 +5,12 @@ import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 
 contract MockUniswapV2Router02 {
     function swapExactTokensForTokens(
-        uint amountIn,
-        uint amountOutMin,
+        uint256 amountIn,
+        uint256 amountOutMin,
         address[] calldata path,
         address to,
-        uint deadline
-    ) external virtual returns (uint[] memory amounts) {
+        uint256 deadline
+    ) external virtual returns (uint256[] memory amounts) {
         IERC20(path[0]).transferFrom(msg.sender, address(this), amountIn);
         IERC20(path[2]).transfer(to, amountIn);
         amounts = new uint256[](path.length);
@@ -18,5 +18,4 @@ contract MockUniswapV2Router02 {
         amounts[1] = amountIn;
         amounts[2] = amountIn;
     }
-
 }
